requirements.txt
openai == 1.26.0
fastapi == 0.111.0
uvicorn == 0.29.0
weaviate-client == 3.26.2
opentelemetry-api == 1.24.0
opentelemetry-sdk == 1.24.0
opentelemetry-exporter-otlp-proto-grpc ==  1.24.0
traceloop-sdk == 0.18.0
langchain_openai == 0.1.7
langchain_community == 0.2.1
beautifulsoup4 == 4.12.3
lxml == 5.2.2
faiss-cpu == 1.8.0
tokencost == 0.1.8